requirements.txt
beautifulsoup4==4.3.2
defusedxml==0.4.1
Django==1.7.3
django-braces==1.4.0
django-crispy-forms==1.4.0
django-webtest==1.7.7
factory-boy==2.4.1
oauthlib==0.7.2
psycopg2==2.5.4
python3-openid==3.0.5
requests==2.5.1
requests-oauthlib==0.4.2
six==1.9.0
waitress==0.8.9
WebOb==1.4
WebTest==2.0.17
